Make openssl we use configurable
