+++ snap/hooks/configure
@@ -267,7 +267,7 @@ then
   # Create the known tokens
   touch ${SNAP_DATA}/credentials/known_tokens.csv
   chmod 660 ${SNAP_DATA}/credentials/known_tokens.csv
-  kubelet_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+  kubelet_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
   hostname=$(hostname)
   echo "${kubelet_token},system:node:${hostname},kubelet-0,\"system:nodes\"" >> ${SNAP_DATA}/credentials/known_tokens.csv
   ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
@@ -292,7 +292,7 @@ then
   # Create the known tokens
   touch ${SNAP_DATA}/credentials/known_tokens.csv
   chmod 660 ${SNAP_DATA}/credentials/known_tokens.csv
-  proxy_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+  proxy_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
   echo "${proxy_token},system:kube-proxy,kube-proxy" >> ${SNAP_DATA}/credentials/known_tokens.csv
 
   ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
@@ -317,7 +317,7 @@ then
   # Create the known tokens
   touch ${SNAP_DATA}/credentials/known_tokens.csv
   chmod 660 ${SNAP_DATA}/credentials/known_tokens.csv
-  scheduler_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+  scheduler_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
   echo "${scheduler_token},system:kube-scheduler,scheduler" >> ${SNAP_DATA}/credentials/known_tokens.csv
   ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
   # Create the client kubeconfig for the scheduler
@@ -342,7 +342,7 @@ then
   # Create the known tokens
   touch ${SNAP_DATA}/credentials/known_tokens.csv
   chmod 660 ${SNAP_DATA}/credentials/known_tokens.csv
-  controller_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+  controller_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
   echo "${controller_token},system:kube-controller-manager,controller" >> ${SNAP_DATA}/credentials/known_tokens.csv
   ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
 
@@ -467,7 +467,7 @@ then
   echo "Removing basic auth"
   # Create the client kubeconfig
   ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
-  admin_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+  admin_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
   echo "${admin_token},admin,admin,\"system:masters\"" >> ${SNAP_DATA}/credentials/known_tokens.csv
   cp ${SNAP}/client.config.template ${SNAP_DATA}/credentials/client.config
   chmod 660 ${SNAP_DATA}/credentials/client.config
