+++ snap/hooks/install
@@ -12,7 +12,6 @@ fi
 ARCH="$(${SNAP}/bin/uname -m)"
 export LD_LIBRARY_PATH="${SNAP}/lib:${SNAP}/usr/lib:${SNAP}/lib/$ARCH-linux-gnu:${SNAP}/usr/lib/$ARCH-linux-gnu"
 export PATH="${SNAP}/usr/sbin:${SNAP}/usr/bin:${SNAP}/sbin:${SNAP}/bin:$PATH:/usr/bin:/usr/local/bin"
-export OPENSSL_CONF="${SNAP}/etc/ssl/openssl.cnf"
 
 cp -r --preserve=mode ${SNAP}/default-args ${SNAP_DATA}/args
 mv ${SNAP_DATA}/args/certs.d/localhost__32000 ${SNAP_DATA}/args/certs.d/localhost:32000
@@ -51,16 +50,16 @@ mkdir -p ${SNAP_DATA}/credentials
 ca_data=$(cat ${SNAP_DATA}/certs/ca.crt | ${SNAP}/usr/bin/base64 -w 0)
 
 # Create the known tokens
-admin_token=$(openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+admin_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
 echo "${admin_token},admin,admin,\"system:masters\"" > ${SNAP_DATA}/credentials/known_tokens.csv
-proxy_token=$(${SNAP}/usr/bin/openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+proxy_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
 echo "${proxy_token},system:kube-proxy,kube-proxy" >> ${SNAP_DATA}/credentials/known_tokens.csv
-kubelet_token=$(${SNAP}/usr/bin/openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+kubelet_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
 hostname=$(hostname)
 echo "${kubelet_token},system:node:${hostname},kubelet-0,\"system:nodes\"" >> ${SNAP_DATA}/credentials/known_tokens.csv
-controller_token=$(${SNAP}/usr/bin/openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+controller_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
 echo "${controller_token},system:kube-controller-manager,controller" >> ${SNAP_DATA}/credentials/known_tokens.csv
-scheduler_token=$(${SNAP}/usr/bin/openssl rand -base64 32 | ${SNAP}/usr/bin/base64)
+scheduler_token=$(openssl_call rand -base64 32 | ${SNAP}/usr/bin/base64)
 echo "${scheduler_token},system:kube-scheduler,scheduler" >> ${SNAP_DATA}/credentials/known_tokens.csv
 
 # Create the client kubeconfig
